Working on Delete

+++ RazorBoatApp2026InClass/Pages/Members/DeleteMember.cshtml.cs
@@ -10,6 +10,7 @@ namespace RazorBoatApp2026InClass.Pages.Members
         private IMemberRepositoryAsync _repo;
 
         public Member? DeleteMember { get; set; }
+        public string Message { get; set; }
         public DeleteMemberModel(IMemberRepositoryAsync repo)
         {
             _repo = repo;
@@ -23,9 +24,21 @@ namespace RazorBoatApp2026InClass.Pages.Members
         public async Task<IActionResult> OnPostDelete(int id)
         {
             Member? m = await _repo.SearchMember(id);
-            
-            await _repo.RemoveMember(m);
-            return RedirectToPage("index");
+            if(m != null)
+            {
+                int? sessionID = HttpContext.Session.GetInt32("ID");
+                int? sessionRoleValue = HttpContext.Session.GetInt32("MemberRole");
+                bool isAdmin = sessionRoleValue.HasValue && (MemberRole)sessionRoleValue == MemberRole.Admin;
+                if (sessionID != m.Id && isAdmin)
+                {
+                    Message = "Du kan ikke ændre denne bruger";
+                    return Page();
+                }
+                await _repo.RemoveMember(m);
+                return RedirectToPage("index");
+            }
+            Message = "Brugeren kunne ikke findes";
+            return Page();
         }
         public IActionResult OnPost()
         {
